refactor: align applist presentation package with its directory
Keep App workspace Compose screens in applist.presentation.
Shared scaffolds stay in ui.compose.
